requirements.txt
# Orchestrator
honcho
watchdog

# Common
pyyaml
attrs
cattrs

# Transnet
zmq
protobuf
protoletariat

# serviz
zmq
flask
flask-socketio
